add 'isVisibleAddMemberText' check in visibilityOptions
